move ManagedGroupDB to dbgroup file, add rebuild method to ManagedGroupDB with documentation to functionality
